requirements.txt
asgiref==3.11.1
Django==6.0.2
django-debug-toolbar==6.2.0
sqlparse==0.5.5
redis
django-redis
gunicorn
psycopg2-binary==2.9.9
stripe 


pytest
pytest-django
pytest-cov
factory-boy